refactor: split napi feature helpers
