+++ crates/ox_content_napi/src/features.rs
@@ -1,6 +1,5 @@
 #![allow(clippy::redundant_pub_crate)]
 
-use compact_str::CompactString;
 use rustc_hash::FxHashMap;
 use std::borrow::Cow;
 use std::path::PathBuf;
@@ -10,13 +9,17 @@ use crate::{
     JsWikiLinkOptions,
 };
 
+mod attr_tokens;
+mod attributes;
 pub(crate) mod code_blocks;
 mod code_imports;
 mod edit;
 mod emoji;
 mod emoji_shortcodes;
+mod segments;
 mod wiki;
 
+use attributes::transform_attribute_syntax;
 pub(crate) use code_blocks::{
     extract_code_blocks, extract_docs_tests, lint_code_blocks, CodeBlockDiagnostic,
     ExtractedCodeBlock,
@@ -24,6 +27,7 @@ pub(crate) use code_blocks::{
 use code_imports::CodeImportOptions;
 use edit::append_edit_this_page;
 use emoji_shortcodes::replace_emoji_shortcodes;
+use segments::transform_markdown_text_segments;
 use wiki::replace_wiki_links;
 
 #[derive(Clone, Default)]
@@ -210,459 +214,6 @@ fn resolve_edit_this_page(
         source_path: source_path.to_string(),
         label: options.label.clone().unwrap_or_else(|| "Edit this page".to_string()),
     })
-}
-
-fn transform_markdown_text_segments(
-    source: &str,
-    mut transform: impl FnMut(&str, &mut String),
-) -> Option<String> {
-    let mut out = String::with_capacity(source.len());
-    let mut changed = false;
-    let mut in_fence = false;
-    let mut fence_char = b'\0';
-    let mut fence_len = 0usize;
-
-    for line_with_end in source.split_inclusive('\n') {
-        let (line, ending) = match line_with_end.strip_suffix('\n') {
-            Some(line) => (line, "\n"),
-            None => (line_with_end, ""),
-        };
-
-        if in_fence {
-            out.push_str(line);
-            out.push_str(ending);
-            if is_closing_fence(line, fence_char, fence_len) {
-                in_fence = false;
-                fence_char = b'\0';
-                fence_len = 0;
-            }
-            continue;
-        }
-
-        if let Some(open) = parse_opening_fence(line) {
-            in_fence = true;
-            fence_char = open.fence_char;
-            fence_len = open.fence_len;
-            out.push_str(line);
-            out.push_str(ending);
-            continue;
-        }
-
-        let before_len = out.len();
-        transform_inline_code_segments(line, &mut out, &mut transform);
-        let appended = &out[before_len..];
-        if appended != line {
-            changed = true;
-        }
-        out.push_str(ending);
-    }
-
-    if changed {
-        Some(out)
-    } else {
-        None
-    }
-}
-
-fn transform_inline_code_segments(
-    line: &str,
-    out: &mut String,
-    transform: &mut impl FnMut(&str, &mut String),
-) {
-    let bytes = line.as_bytes();
-    let mut cursor = 0usize;
-    while cursor < bytes.len() {
-        let Some(relative) = memchr::memchr(b'`', &bytes[cursor..]) else {
-            transform(&line[cursor..], out);
-            return;
-        };
-        let tick_start = cursor + relative;
-        transform(&line[cursor..tick_start], out);
-        let tick_count = count_repeated_byte(bytes, tick_start, b'`');
-        let code_start = tick_start + tick_count;
-        if let Some(close) = find_closing_backticks(bytes, code_start, tick_count) {
-            out.push_str(&line[tick_start..close + tick_count]);
-            cursor = close + tick_count;
-        } else {
-            out.push_str(&line[tick_start..]);
-            return;
-        }
-    }
-}
-
-fn transform_attribute_syntax(html: &str) -> Option<String> {
-    let bytes = html.as_bytes();
-    let mut out = String::with_capacity(html.len());
-    let mut cursor = 0usize;
-    let mut changed = false;
-
-    while cursor < bytes.len() {
-        let Some(relative) = memchr::memchr(b'{', &bytes[cursor..]) else {
-            break;
-        };
-        let attr_start = cursor + relative;
-        let Some(attr_end) = find_attr_block_end(html, attr_start) else {
-            cursor = attr_start + 1;
-            continue;
-        };
-        let attrs = &html[attr_start + 1..attr_end];
-        let Some(parsed) = ParsedAttrs::parse(attrs) else {
-            cursor = attr_start + 1;
-            continue;
-        };
-
-        if try_apply_attrs_inside_element(html, &mut out, cursor, attr_start, attr_end, &parsed) {
-            changed = true;
-            cursor = attr_end + 1;
-            continue;
-        }
-
-        if try_apply_attrs_to_previous_element(html, &mut out, cursor, attr_start, &parsed)
-            .is_some()
-        {
-            changed = true;
-            cursor = attr_end + 1;
-            continue;
-        }
-        cursor = attr_start + 1;
-    }
-
-    if !changed {
-        return None;
-    }
-    out.push_str(&html[cursor..]);
-    Some(out)
-}
-
-fn try_apply_attrs_inside_element(
-    html: &str,
-    out: &mut String,
-    cursor: usize,
-    attr_start: usize,
-    attr_end: usize,
-    attrs: &ParsedAttrs,
-) -> bool {
-    let close_start = attr_end + 1;
-    if !html[close_start..].starts_with("</") {
-        return false;
-    }
-    let Some(close_name_end) = html[close_start + 2..].find('>') else {
-        return false;
-    };
-    let tag_name =
-        html[close_start + 2..close_start + 2 + close_name_end].trim().to_ascii_lowercase();
-    if !is_attr_target_tag(&tag_name) {
-        return false;
-    }
-    let open_marker = format!("<{tag_name}");
-    let Some(open_start) = html[..attr_start].rfind(&open_marker) else {
-        return false;
-    };
-    let Some(open_end) = scan_tag_end(html, open_start) else {
-        return false;
-    };
-    if open_end > attr_start {
-        return false;
-    }
-
-    let text_end = html[..attr_start].trim_end().len();
-    out.push_str(&html[cursor..open_start]);
-    out.push_str(&html[open_start..open_end - 1]);
-    write_attrs(out, attrs);
-    out.push('>');
-    out.push_str(&html[open_end..text_end]);
-    true
-}
-
-fn try_apply_attrs_to_previous_element(
-    html: &str,
-    out: &mut String,
-    cursor: usize,
-    attr_start: usize,
-    attrs: &ParsedAttrs,
-) -> Option<usize> {
-    let before = &html[..attr_start];
-    let trimmed_end = before.trim_end().len();
-    if trimmed_end == 0 || trimmed_end > attr_start {
-        return None;
-    }
-    let whitespace = &html[trimmed_end..attr_start];
-
-    if let Some((tag_start, tag_end, close_end)) = find_previous_wrapped_element(html, trimmed_end)
-    {
-        out.push_str(&html[cursor..tag_start]);
-        out.push_str(&html[tag_start..tag_end]);
-        write_attrs(out, attrs);
-        out.push_str(&html[tag_end..trimmed_end]);
-        out.push_str(whitespace);
-        return Some(close_end);
-    }
-
-    if let Some((tag_start, tag_end)) = find_previous_void_element(html, trimmed_end) {
-        out.push_str(&html[cursor..tag_start]);
-        out.push_str(&html[tag_start..tag_end]);
-        write_attrs(out, attrs);
-        out.push_str(&html[tag_end..trimmed_end]);
-        out.push_str(whitespace);
-        return Some(tag_end);
-    }
-
-    None
-}
-
-fn find_previous_wrapped_element(html: &str, end: usize) -> Option<(usize, usize, usize)> {
-    let close_start = html[..end].rfind("</")?;
-    if close_start + 2 >= end {
-        return None;
-    }
-    let close_name_end = html[close_start + 2..end].find('>')? + close_start + 2;
-    let tag_name = html[close_start + 2..close_name_end].trim().to_ascii_lowercase();
-    if tag_name.is_empty() || !is_attr_target_tag(&tag_name) {
-        return None;
-    }
-    let open_marker = format!("<{tag_name}");
-    let open_start = html[..close_start].rfind(&open_marker)?;
-    let open_end = scan_tag_end(html, open_start)?;
-    Some((open_start, open_end - 1, close_name_end + 1))
-}
-
-fn find_previous_void_element(html: &str, end: usize) -> Option<(usize, usize)> {
-    let open_start = html[..end].rfind('<')?;
-    let open_end = scan_tag_end(html, open_start)?;
-    if open_end != end {
-        return None;
-    }
-    let name_start = open_start + 1;
-    let mut name_end = name_start;
-    let bytes = html.as_bytes();
-    while name_end < bytes.len()
-        && !bytes[name_end].is_ascii_whitespace()
-        && bytes[name_end] != b'>'
-        && bytes[name_end] != b'/'
-    {
-        name_end += 1;
-    }
-    let tag_name = html[name_start..name_end].to_ascii_lowercase();
-    if matches!(tag_name.as_str(), "img" | "br" | "hr" | "input") {
-        Some((open_start, open_end - 1))
-    } else {
-        None
-    }
-}
-
-fn is_attr_target_tag(tag: &str) -> bool {
-    matches!(
-        tag,
-        "h1" | "h2"
-            | "h3"
-            | "h4"
-            | "h5"
-            | "h6"
-            | "a"
-            | "img"
-            | "code"
-            | "pre"
-            | "p"
-            | "div"
-            | "span"
-            | "blockquote"
-            | "table"
-            | "tr"
-            | "th"
-            | "td"
-            | "ul"
-            | "ol"
-            | "li"
-    )
-}
-
-fn scan_tag_end(html: &str, start: usize) -> Option<usize> {
-    let bytes = html.as_bytes();
-    let mut i = start;
-    let mut quote = None;
-    while i < bytes.len() {
-        match quote {
-            Some(q) if bytes[i] == q => quote = None,
-            Some(_) => {}
-            None if bytes[i] == b'"' || bytes[i] == b'\'' => quote = Some(bytes[i]),
-            None if bytes[i] == b'>' => return Some(i + 1),
-            None => {}
-        }
-        i += 1;
-    }
-    None
-}
-
-fn find_attr_block_end(html: &str, start: usize) -> Option<usize> {
-    let bytes = html.as_bytes();
-    let mut i = start + 1;
-    while i < bytes.len() {
-        match bytes[i] {
-            b'}' => return Some(i),
-            b'\n' | b'\r' | b'<' | b'>' => return None,
-            _ => i += 1,
-        }
-    }
-    None
-}
-
-#[derive(Default)]
-struct ParsedAttrs {
-    id: Option<String>,
-    classes: Vec<String>,
-    attrs: Vec<(String, String)>,
-}
-
-impl ParsedAttrs {
-    fn parse(value: &str) -> Option<Self> {
-        if !value.contains('#') && !value.contains('.') && !value.contains('=') {
-            return None;
-        }
-        let mut parsed = Self::default();
-        for token in split_attr_tokens(value) {
-            if let Some(id) = token.strip_prefix('#') {
-                if !id.is_empty() {
-                    parsed.id = Some(id.to_string());
-                }
-            } else if let Some(class) = token.strip_prefix('.') {
-                if !class.is_empty() {
-                    parsed.classes.push(class.to_string());
-                }
-            } else if let Some((name, raw_value)) = token.split_once('=') {
-                let name = name.trim();
-                if is_safe_attr_name(name) {
-                    parsed.attrs.push((
-                        name.to_string(),
-                        raw_value.trim_matches(|ch| ch == '"' || ch == '\'').to_string(),
-                    ));
-                }
-            }
-        }
-        if parsed.id.is_none() && parsed.classes.is_empty() && parsed.attrs.is_empty() {
-            None
-        } else {
-            Some(parsed)
-        }
-    }
-}
-
-fn split_attr_tokens(value: &str) -> Vec<&str> {
-    let bytes = value.as_bytes();
-    let mut tokens = Vec::new();
-    let mut cursor = 0usize;
-    while cursor < bytes.len() {
-        while cursor < bytes.len() && bytes[cursor].is_ascii_whitespace() {
-            cursor += 1;
-        }
-        if cursor >= bytes.len() {
-            break;
-        }
-        let start = cursor;
-        let mut quote = None;
-        while cursor < bytes.len() {
-            match quote {
-                Some(q) if bytes[cursor] == q => quote = None,
-                Some(_) => {}
-                None if bytes[cursor] == b'"' || bytes[cursor] == b'\'' => {
-                    quote = Some(bytes[cursor]);
-                }
-                None if bytes[cursor].is_ascii_whitespace() => break,
-                None => {}
-            }
-            cursor += 1;
-        }
-        tokens.push(&value[start..cursor]);
-    }
-    tokens
-}
-
-fn write_attrs(out: &mut String, attrs: &ParsedAttrs) {
-    if let Some(id) = &attrs.id {
-        out.push_str(" id=\"");
-        escape_html_attr(id, out);
-        out.push('"');
-    }
-    if !attrs.classes.is_empty() {
-        out.push_str(" class=\"");
-        for (index, class) in attrs.classes.iter().enumerate() {
-            if index > 0 {
-                out.push(' ');
-            }
-            escape_html_attr(class, out);
-        }
-        out.push('"');
-    }
-    for (name, value) in &attrs.attrs {
-        out.push(' ');
-        out.push_str(name);
-        out.push_str("=\"");
-        escape_html_attr(value, out);
-        out.push('"');
-    }
-}
-
-struct FenceOpen {
-    fence_char: u8,
-    fence_len: usize,
-    language: CompactString,
-    meta: CompactString,
-}
-
-fn parse_opening_fence(line: &str) -> Option<FenceOpen> {
-    let trimmed = line.trim_start();
-    let bytes = trimmed.as_bytes();
-    let fence_char = *bytes.first()?;
-    if fence_char != b'`' && fence_char != b'~' {
-        return None;
-    }
-    let fence_len = count_repeated_byte(bytes, 0, fence_char);
-    if fence_len < 3 {
-        return None;
-    }
-    let rest = trimmed[fence_len..].trim();
-    let mut parts = rest.splitn(2, char::is_whitespace);
-    let language = CompactString::from(parts.next().unwrap_or_default());
-    let meta = CompactString::from(parts.next().unwrap_or_default().trim());
-    Some(FenceOpen { fence_char, fence_len, language, meta })
-}
-
-fn is_closing_fence(line: &str, fence_char: u8, fence_len: usize) -> bool {
-    let trimmed = line.trim();
-    let bytes = trimmed.as_bytes();
-    bytes.len() >= fence_len
-        && bytes[..fence_len].iter().all(|value| *value == fence_char)
-        && bytes[fence_len..].iter().all(|value| *value == fence_char)
-}
-
-fn count_repeated_byte(bytes: &[u8], start: usize, byte: u8) -> usize {
-    let mut count = 0usize;
-    let mut cursor = start;
-    while cursor < bytes.len() && bytes[cursor] == byte {
-        count += 1;
-        cursor += 1;
-    }
-    count
-}
-
-fn find_closing_backticks(bytes: &[u8], from: usize, count: usize) -> Option<usize> {
-    let mut cursor = from;
-    while cursor < bytes.len() {
-        let relative = memchr::memchr(b'`', &bytes[cursor..])?;
-        let start = cursor + relative;
-        if count_repeated_byte(bytes, start, b'`') >= count {
-            return Some(start);
-        }
-        cursor = start + 1;
-    }
-    None
-}
-
-fn is_safe_attr_name(name: &str) -> bool {
-    !name.is_empty()
-        && !name.to_ascii_lowercase().starts_with("on")
-        && name
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b':' | b'_'))
 }
 
 fn escape_html_text(value: &str, out: &mut String) {
